Added a contenttype field to "Contenido"
This allows to create a general content report using only the
Contenido table, and makes all inheritance usage easier.

Added a summary report and view

Using Dynpath and the new conttentypes from Contenido we can
stablish the a reports of the content that was requested in
a given period. the sms_report template was modified to add a
temporary ugly table as proof on concept